Export performance improvement.
Previously, the export mapper would convert a Text writable
representing a record into a SqoopRecord, which would then again
be converted back to String representation. This resulted in load
speeds of 13 MB/sec. This change alters this behavior to bypass
the conversion to SqoopRecord resulting in load speed up to
110 MB/sec.